Split api module functions into encode and compare submodules
