+++ kissim/api/__init__.py
@@ -1,0 +1,6 @@
+"""
+kissim.api
+"""
+
+from .encode import encode
+from .compare import compare
