+++ kissim/api/encode.py
@@ -1,7 +1,7 @@
 """
-kissim.api
+kissim.api.encode
 
-Main API for kissim encoding and comparison.
+Main API for kissim encoding.
 """
 
 import logging
@@ -9,7 +9,6 @@ import logging
 from opencadd.databases.klifs import setup_remote, setup_local
 
 from kissim.encoding import FingerprintGenerator
-from kissim.comparison import FeatureDistancesGenerator, FingerprintDistanceGenerator
 
 logger = logging.getLogger(__name__)
 
@@ -50,43 +49,6 @@ def encode(structure_klifs_ids, json_filepath=None, n_cores=1, local_klifs_sessi
         fingerprints.to_json(json_filepath)
 
     return fingerprints
-
-
-def compare(
-    fingerprint_generator,
-    csv_path=None,
-    n_cores=1,
-    distance_measure="scaled_euclidean",
-    feature_weights="101",
-):
-    """
-    Compare fingerprints (pairwise).
-
-    Parameters
-    ----------
-    fingerprint_generator : kissim.encoding.FingerprintGenerator
-        Fingerprints for KLIFS dataset.
-    csv_path : str
-        TODO
-    n_cores : int
-        Number of cores used to generate fingerprint distances.
-    distance_measures : str
-        Distance measures TODO.
-    feature_weights : str
-        Feature weighting scheme.
-    """
-
-    print(csv_path)
-
-    feature_distances_generator = FeatureDistancesGenerator()
-    feature_distances_generator.from_fingerprint_generator(fingerprint_generator, distance_measure)
-    # TODO save to file
-
-    fingerprint_distance_generator = FingerprintDistanceGenerator()
-    fingerprint_distance_generator.from_feature_distances_generator(
-        feature_distances_generator, feature_weights
-    )
-    # TODO save to file
 
 
 def _setup_klifs_session(local_klifs_session=None):
